% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{columns}
\alias{columns}
\title{Filter Columns from Data Frame, Matrix, or List}
\usage{
columns(out, cols = NULL)
}
\arguments{
\item{out}{A data frame, matrix, atomic vector, or a list of such objects.}

\item{cols}{A numeric vector of column indices or a character vector of column names to keep.}
}
\value{
A filtered object with only the specified columns. The output will match the input type (matrix, data frame, or list).
}
\description{
Filters specified columns from a data frame, matrix, or list of such objects.
}
\examples{
df <- data.frame(Estimate = 1:3, SE = 0.1, Extra = 4:6)
columns(df, cols = c("Estimate", "SE"))

}
