% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVISE-package.R
\docType{package}
\name{DEVISE}
\alias{DEVISE}
\alias{DEVISE-package}
\title{DEVISE: Data Exploration, Visualization, and Inference with Statistical Estimation}
\description{
`DEVISE` is a companion toolkit for working with other packages that focus on statistical estimation. 
Where other packages provide high-accuracy confidence intervals for a wide variety of statistics, 
`DEVISE` handles the pre-processing of data (extraction of data, calculation of descriptives) and 
post-processing of results into more accessible and readable formats (such as summary tables and comparison plots) 
all using a pipe-friendly framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwendorf/DEVISE}
  \item Report bugs at \url{https://github.com/cwendorf/DEVISE/issues}
}

}
\author{
\strong{Maintainer}: Craig Wendorf
\email{cwendorf@uwsp.edu}
}
