% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{intervals}
\alias{intervals}
\title{Extract Point Estimate and Confidence Interval Columns}
\usage{
intervals(x)
}
\arguments{
\item{x}{A data frame, matrix, or list with \code{estimate} and \code{interval} elements.}
}
\value{
A data frame or matrix with columns: Estimate, LL, UL.
}
\description{
Selects the point estimate and confidence interval columns from a data frame, matrix, or list.
}
